Add Puppetfile.tempest file
Adds Puppetfile.tempest file that serves as a
specification for all modules that need to be
installed for tempest.

This is specified in its own file so that it
does not have to download every openstack module
just to build tempest.

This decision may not be the correct desicion b/c
it means that we have to maintain versions of the
same modules in two Puppetfiles (stdlib, ...)