Fix ip address code
Previously, the ip address code did not
necessarily get the correct port to associate
the address to.

This is b/c it was only searching for the ip address
of the port and not the subnet. In the case of multiple
stacks, it is possible that there will be multiple VMs
active with the same io address.

This commit updates the code that finds the port to attach
the ip address to, and ensures that the specified port
if the correct on (based on both subnet id and address).

It also updates the floating ip deletion code.